PROJECT-12-T-12: rename control package
