+++ src/main/java/local/uniclog/MainAppUi.java
@@ -5,7 +5,7 @@ import javafx.fxml.FXMLLoader;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
 import javafx.stage.Stage;
-import local.uniclog.ui.controlls.controls.SceneControlService;
+import local.uniclog.ui.controlls.actions.SceneControlService;
 
 import java.io.IOException;
