+++ src/main/java/local/uniclog/ui/AppController.java
@@ -9,9 +9,9 @@ import local.uniclog.model.actions.impl.Sleep;
 import local.uniclog.model.actions.types.ActionType;
 import local.uniclog.model.actions.types.EventStateType;
 import local.uniclog.model.actions.types.MouseButtonType;
-import local.uniclog.ui.controlls.controls.ControlServiceAbstract;
-import local.uniclog.ui.controlls.controls.SceneControlService;
-import local.uniclog.ui.controlls.controls.impl.*;
+import local.uniclog.ui.controlls.actions.ControlServiceAbstract;
+import local.uniclog.ui.controlls.actions.SceneControlService;
+import local.uniclog.ui.controlls.actions.impl.*;
 import local.uniclog.ui.controlls.model.ControlPack;
 import local.uniclog.utils.DataUtils;
 import lombok.extern.slf4j.Slf4j;
