+++ src/main/java/local/uniclog/ui/controlls/actions/ControlServiceAbstract.java
@@ -1,4 +1,4 @@
-package local.uniclog.ui.controlls.controls;
+package local.uniclog.ui.controlls.actions;
 
 import javafx.application.Platform;
 import local.uniclog.model.actions.ActionsInterface;
