+++ src/main/java/local/uniclog/ui/controlls/actions/impl/DefaultControl.java
@@ -1,6 +1,6 @@
-package local.uniclog.ui.controlls.controls.impl;
+package local.uniclog.ui.controlls.actions.impl;
 
-import local.uniclog.ui.controlls.controls.ControlServiceAbstract;
+import local.uniclog.ui.controlls.actions.ControlServiceAbstract;
 
 import static java.util.Objects.isNull;
 import static local.uniclog.utils.ConfigConstants.TEMPLATE_NOT_SET_CONTROLS;
